resume_ats_backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
langchain-google-genai
PyPDF2
